Fix test data for call hierarchy test
New type inference method allows to resolve parameters of 'C.func1' and 'bar1' to
class A. Because of that these functions are added in results of findUsages during
later pass when we check that references to 'target_func' found inside their body
may be resolved to A.target_func (not to B.target_func as in first pass).
Hence the updated order.