[push] IDEA-131438 IDEA-130650 Make remote selector independent from edit mode, and notify model about remote change
Introduce a special tag which will be called on click with a mouse
event which produced that click. Use this event coordinates to show
the remote selection popup near the click area.

Add a callback to the PushTargetPanels which should be called when
the target panel changes the model not via the usual edit mode.