[git] Fix non-ASCII paths unescaping problems
Fix GitUtil.unescapePath as proposed in PR#123: i is incremented in
the inner loop even after the octal part finishes => it should be
decremented back, so that next character is not missed. Add a test.
(https://github.com/JetBrains/intellij-community/pull/123)

However, it doesn't fix problems with unicode paths containing
diacritic symbols (in spite of PR#218 which proposes fix for
IDEA-132150). Such paths are always returned by Git in
octally-escaped NFD (decomposed) form, while on disk and/or in the
repository they can be stored in NFC or even non-normalized form.

Therefore make Git return unescaped paths by passing
-c core.quotepath=false to all commands.
This option is not supported for Gits older than 1.7.2 - they will still
receive escaped paths and use patched unescapePath(). Others will
receive those characters which are stored in Git database.

This won't fix the situation if a non-normalized path is stored to
the repository on Linux/Windows and then checked out on Mac (which
HFS+ automatically normalizes to NFC on read/write): the path on disk
won't match the path in repository.

But for other cases this fixes the following issues:
IDEA-84715, IDEA-116247, IDEA-132150