diff: fix scrolling position restore
The issue was with collapsed foldings. The position is captured with folded editors, while restoring was in onInit() - before than foldings are initialized.
Solution: wait until onRediff() or onSlowRediff() to apply scrolling position. In case of onSlowRediff() result will be the same, but it's unlikely behaviour. And it will happen when document is changed, so exact position restore is problematic.