IDEA-135522 First click to gutter after popup menu is not handled
Remove the flag myPopupInvokedOnPress, since it is effective not
for direct mousePressed-mouseReleased cycle, but stays for the next
mouseRelease as well, thus preventing the click to be handled.

The reason why mouseReleased is not called for the right click -
because the event is consumed in invokePopup. That also fixes the
issue, probably for which the flag was introduced.