PY-12749 Sparingly use continuation indents for arguments in the headers of control statements
Usage of continuation indents (which presumably are wider than normal
indents) in the headers of control statements was supposed to prevent
PEP8's error E125 "Continuation line with same indent as next logical
line". However such indents is not necessary for e.g. nested function
calls where indent of the arguments is already visually distinctive
(larger) from the offset of the following statement list. To
distinguish such cases I've added check that argument list is located
on the same line as the header of the containing control statement. If
line numbers differ we use normal indent for the arguments and
continuation indent otherwise. It doesn't matter for complex
non-hanging indent cases like if foo(bar(1... because alignment will
be used there anyway.