Remove formatter logic that prevented proper alignment of empty collection literals
It was added a while back for PY-6751, but corresponding test
passes without it now. Also it turned out that
PyIndentTest#testEnterInDictInTuple depended on this change, but
behavior tested there was recreated (once again) by proper use of
hasHangingIndent in formatter, i.e. once opening and closing braces of
the dict are on different lines, the literal should not have alignment.