[vcs] Refactor ChangeListsIndexes: use FilePath instead of absolute path
Seems to be useful on its own (working with absolute paths looks rather
strange), but is also needed for correct fix of IDEA-137136.

Deprecate ChangeListManagerGate#getFileStatus(File) and overload it
for the FilePath instead. Wrap File to FilePathImpl in the deprecated
method.

Use correct FilePathImpl comparator for the map, to fix haveChangesUnder