initial fix for numerous bugs mostly related to to multiline method call parameters. We want to ensure that in sequence: "Parenthesis -> Method Params -> Method Params Children", each member has less indent than it's right sibling.
Currently we check if any param is located on new line, and if so, we enforce children indent. Additionally we check if parenthesis has the same indent as any children, indent is enforced also.

This checks and modifications are performed after formatter's AdjustWhiteSpaceState, when we know all proper whitespaces states,  but before actual modification.