PY-14749 Add different types for Python remote mappings to treat them according to their natural priorities
User-defined mappings comes first, then Vagrant synced folders and SSH deployment mappings, then sys.path, helpers and skeletons. Adding mapping priorities should fix the issue when there are several mappings that suits for the remote path and we are choosing not the best one (e.g. when one mapping is from sys.path and the other is Vagrant synced folder).