PyArgumentsListInspection now uses new PyCallExpressionHelper.mapArguments() instead of CallArgumentsMapping
We simplified two checks: now we don't detect some cases of
multiple values that resolve to a single positional parameter (we show
a related warning about the wrong order of keyword arguments anyway)
and we stopped telling the user that a sequence literal argument doens't
match against a tuple paramater in Python 2. We might enhance the new
version of arguments-parameters matching to restore this functionality.