diff: remove O(N*M) complexity in LineFragmentSplitter
where N is length of the text, and M is number of lines

The issue was with very long sequence of ignore-whitespace-equal lines.

All words inside are equal, so we call 'addLineChunk()' once for each line - O(M).
These blocks 'hasEqualWords' inside and are 'equalsIgnoreWhitespace'. Thus on each addLineChunk() we merged new block with previous one.
To do so, we checked last block on isEqualsIgnoreWhitespace() - O(N).

the fix is to compute 'isEqualIgnoreWhitespaces()' once for each new block, without recomputing it on each merge.