Take compatible plugin instead of incompatible even if the newest has less version
Newer plugin can have version that less then older one in another branch. If IDE was updated to another branch, plugin becomes incompatible and it should be updated to the newer compatible one even if it has less version.

In general avoid having incompatible plugin when there is a newer one that is compatible is a good thing.