[java] using "InnerClasses" attribute for resolving binary names in .class stubs (IDEA-82515)
Previously, IDEA relied on heuristics to resolve binary names like "Map$Entry" into a source form ("Map.Entry"). The problem with this approach is that '$' is a legal part of Java identifiers. The fix (1) consolidates the heuristics from two different places into a single mapper, and (2) allows to use "InnerTables" attribute instead (according to JVMS 4.7.6, it should contain at least some of the referenced inner classes) - this part is disabled by default.