Reset tree element pointer only when the document has been commited (fixes blinking test)
This fixes a non-determistic bug when the document commit thread
is loading the AST via PsiFile.getNode() during the commit (new behavior
in 94863a8df423d3568167f709371d42c7410d0946) and thus setting a non-null
value to the tree element pointer while the AWT thread has already set
the tree element pointer to be null. The AWT thread in this test should
be the last one to reset the tree element pointer.