Make NativeFileWatcherImpl#isRepetition protected.
Some file systems aren't 100% accurate in their VFS events,
causing missed events. For instance, consider for a single file:

CHANGE - DELETE - CREATE - CHANGE

If the create is missed, then the file won't be updated because
it looks like a repetition.

This allows implementors for those file systems to override this
and get the CHANGE event.