git: report if both current revision and current branch are null
This is not a valid situation, which indicates some issue
while reading .git/HEAD and .git/refs.

E.g. current branch is null in detached HEAD, but then there is
current revision. On the other hand, current revision is null in a
fresh repository, which is however on branch.

Relates to IDEA-101222.