cleanup: make VcsDirectoryMapping less mutable & less nullable
VcsMappingConfigurationDialog used to modify existing mapping objects,
which are contained in the mappings table model.
Now it substitutes the values the table with new values instead.

This allows to remove remove VcsDirectoryMapping() empty
constructor which introduces unnecessary nullability to the
mapping fields, and also make myDirectory final since it is not changes
anywhere else.