Currently, we call getState after loadState immediately on init component persistence state.
As a side effect, it also checks that you implement loadState correctly. Because if not — you will get error "Content equals, but it must be handled not on this level: misc.xml" (only in test mode). Yeah — this assertion error is not clear and it is not easy to understand on which level the error is.