git: get rid of GitPlatformFacade as not valuable anymore
The interface was introduce as an attempt to separate git (dvcs) module
from the rest of the platform, to be able to write module tests
(faster and more reliable than the full-platform one).

However, the idea didn't proove itself: it appeared to be too tedious
to substitute all platform references by the facade methods, while
the platform tests became faster and used now pretty well.

Substitute the usages by actual calls to the platform methods.
Remove GPF usages and methods, or deprecate those which are called
by 3-party plugins.