BinaryOutputReader should support readAvailableBlocking (as well as all other BaseDataReader descendants)
When BinaryOSProcessHandler is executed in PTY mode it assumes that readers read the data in blocking mode, while not all readers support it, which results into premature stream close.
As it is not possible to mark readAvailable as final due to compatibility (as well as having readAvailableBlocking and readAvailableNonBlocking to be abstract), javadoc is added and the default implementations raise NotSupportedOperationException.