IDEA-CR-10973 Don't lose ItemEvent for SDK created in PythonSdkChooserCombo
It might happen because ProjectSdkModel copies added SDK internally
and that original SDK is no longer in the list returned by
PyConfigurableInterpreterList which is then used to initialize
the model of the combo box in the project wizard.
As a result passing that SDK to setSelectedItem() has no effect,
SDK doesn't appear in the combo box and checkValid() isn't invoked
either.