IDEA-128727 Uncommit: git reset HEAD^ from the log at the last commit
Restrictions:
* Allowed only for the last commit in the current branch.
* Not allowed for commit pushed to a protected branch.
* Not allowed for merge commits.

Checking for containing branches can be slow, therefore in update()
only cache information is used if available. actionPerformed() gets
the actual unformation under a modal task if needed.
A handy method returning Future was added to ContainingBranchesGetter
to guarantee that the request will finish.

Prefer VcsShortCommitDetails here, because we need parents and subject,
while we can load changes later. On the other hand, using
VcsFullCommitDetails may not work since they are loaded later after
user selects a commit. We need the changes from VcsFullCommitDetails
when applying reset command, but the loading is performed in a bg task
then.