PY-20058 Suppress E402 errors from PEP 8 inspection in IPython notebook cells
The error is suppressed only when magic commands cause it, not any
other unexpected statement preceding an import. It doesn't work for
docstrings preceded by magic commands since our lexer cannot recognize
such dangling string literals as docstrings in the first place, but it
seems out of the scope for this issue.