PY-18816 Remove failing tests from PyStubsTest since they check not existing functionality
We decided to broaden the scope of PY-18816 so that Python 3-style
function annotations (PEP 3107) are persisted in stub files as well.
Since it requires much more efforts and won't be implemented soon it
doesn't make much sense to keep these tests in the trunk. Dedicated
feature branch contains them anyway.