PY-20071 Fixed: Nonexistent logger methods should be mark as unresolved PY-19716 Fixed: logger.exception() without arguments doesn't show a warning
Skeleton for logging.getLogger was accidentally pushed to primary python-skeletons repo and then was synced with current repo.
So there is no skeleton in affected files but tests are based on it.

At first logger.getLogger returned union of Logger and RootLogger.
RootLogger just inherits Logger and doesn't introduce any additional functionality.
But there was a problem with PyArgumentListInspection which doesn't work in case of many callees (e.g. logger.exception).
So skeleton was updated to return only logger from logging.getLogger.