Reset __PYVENV_LAUNCHER__ set by MacPorts that breaks site-packages in venvs (PY-18074)
Python 3.3+ interpreters installed via MacPorts set the system-wide
__PYVENV_LAUNCHER__ environment variable to point to the original
Python interpreter. That prevents any venvs from setting their correct
site-packages directories, because this variable is used in site.py
initialization logic for setting an alternative root for venvs.

This fix sets the __PYVENV_LAUNCHER__ to point to the correct location
of the project intepreter if this environment variable exists.