file history: don't force refresh if data is taken from cache
The UI is refreshed anyway in finished(),
while forceRefresh() is going to schedule yet another data request.
The data will be taken from cache again, thus no additional requests
to the VCS, but it will cause another forceRefresh().
As a result, it used to cause unnecessary chained later UI repaints,
but since many invokeLater-s became invokeAndWait, it leads to SOE.

This also partially fixes IDEA-141802: when data is cached,
but not in the case of full refresh.