IDEA-141802 file history: don't blink with empty table until the data is loaded
reportCreatedEmptySession is named incorrectly, since it is used
to report not only empty history sessions, but other ones as well.

Moreover, any reportCreatedEmptySession led to the file history UI refresh.
It is correct for non-empty sessions, but not needed for empty sessions
which actually indicate just a start of the history loading process.

This fix would be incorrect if a valid history with no revisions could
exist, but since it is impossible, any empty history means "not loaded yet".