IDEA-159580 fix ChangeListManager.freeze() method which hangs sometimes.
Semaphore should be turned down() before up(),
otherwise it may waitFor().

Using standard Semaphore instead of SemaphoreContinuationContext
since the architecture of the latter doesn't differ down() and waitFor().

In addition to that, make the waitFor() method cancellable.

This fix is for branch 162.
The deprecated API is not used externally and will be removed in master.