Intercept duplicate Languages at init and suggest disabling responsible plugin(s)
This commit introduces ImplementationConflictException which is handled if logged as error.

On handling, implementation classes are mapped to corresponding plugins. If there is at least
one implementation class located in platform jars, the others have to be disabled. In other
case, it is possible to leave one plugin (and thus its implementation) alive and try to restart.