PY-20776 PY-20775 PY-20773 Do not highlight empty ranges inside f-strings
Such errors are difficult to spot because most of the time the
corresponding markers are displayed only in the right gutter, not in
the editor (though, sometimes a single character is highlighted for
some reason).

For an empty expression fragment the whole its content range is
highlighted (from "{" to either "!", ":" or "}").

For the first unclosed expression fragment the range from "{" to the
end of the containing string node is highlighted.

In case of a missing conversion character the character "!" itself is
highlighted.